ELECTRON-602: fix the pre-install script on macOS installation
Currently, when we try to install a version that is lower than the currently installed version, the pre install script deletes the existing version and the new version is not installed either. This is because of how macOS handles upgrades.

So, we add a version check function that exits the script if we have a new version already installed on the system rather than deleting the existing version installed on the system