ELECTRON-805: fix protocol handler uninstall issues on Windows (#510)
Currently, when we uninstall symphony on Windows, the protocol handler registry entry is left as is because it is created by Electron in the first place.

To avoid such leftovers, we add logic to use Electron to register protocol handler only on Mac. On Windows, we use the installer to create a registry entry and manage it during installation as well.