ELECTRON-1138: support opening email composition in new window (#601)
* bump up version number to 3.7.2

* ELECTRON-1138: support opening email composition in new window

Currently, due to security reasons, we don't open any url which is not the same as the main window url in a new window, instead, we open it in a new tab of the system default browser.

However, this impacts the email app as we'll need new window to be opened to compose an email. This commit fixes that issue by checking if we need to open a sub-domain of the pod url domain.