SDA-1525: add support for pod whitelisting (#787)
Currently, any url passed via command line is not validated, and, SDA essentially acts as a browser. To counter this, we add a "podWhitelist" parameter in the config file that allows admins to configure what pod urls can be passed via command line. And, only those URLs will be supported to be loaded via command line